backend/codingpup/requirements.txt
asgiref==3.6.0
Django==4.2
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
ez-setup==0.9
phonenumbers==8.13.8
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
